+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## -=[ Version 21 - 02/14/2026]=-
 
 - ADD mouse wheel support for Speed and Scale sliders on Main tab
+- ADD mouse wheel support for Transparency sliders on Misc tab
 
 - REPLACE Hover-mode with Spectator mode on Main tab
   - Replaced non-functional Hover-mode buttons with Spectator mode (.gm spectator on/off)